requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.6.1

# GitHub Integration
PyGithub==2.1.1

# AI Integration
google-genai==1.55.0

# GCP
google-cloud-secret-manager==2.16.4

# HTTP Requests
requests==2.31.0

# Testing (optional for now)
pytest==7.4.3
pytest-asyncio==0.21.1